backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
requests==2.31.0
resend==0.5.2 